pyproject.toml
[tool.poetry]
name = "dewarp"
version = "0.1.0"
description = "A document scanning and dewarping tool that corrects perspective distortion"
authors = ["Stefan Janse van Rensburg <stefanj@mandela.ac.za>"]
readme = "README.md"
packages = [{include = "dewarp", from = "src"}]
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Image Processing",
]
include = [
    "src/dewarp/resources/weights/*.pth"
]

[tool.poetry.dependencies]
python = "^3.12.3"
torch = "^2.0.0"
numpy = "^2.1.2"
opencv-python = "^4.10.0.84"
pillow = "^11.0.0"
tqdm = "^4.65.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-sugar = "^1.0.0"
black = "^24.10.0"
isort = "^5.12.0"
flake8 = "^7.1.1"
mypy = "^1.5.0"

[tool.poetry.scripts]
dewarp = "dewarp.cli.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=dewarp --cov-report=term-missing"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["src/dewarp"]
omit = [
    "tests/*",
    "**/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
show_missing = true
fail_under = 80